+++ x/gov/bucket_test.go
@@ -171,27 +171,27 @@ func TestQueryElectorate(t *testing.T) {
 	}{
 		"By ID and first version": {
 			path: "/electorates",
-			data: mustMarshal(t, vRefV1),
+			data: orm.MarshalVersionedID(*vRefV1),
 			exp:  []Electorate{electorateV1},
 		},
 		"By ID and second version": {
 			path: "/electorates",
-			data: mustMarshal(t, vRefV2),
+			data: orm.MarshalVersionedID(*vRefV2),
 			exp:  []Electorate{electorateV2},
 		},
 		"By ID and unknown version": {
 			path: "/electorates",
-			data: mustMarshal(t, &orm.VersionedIDRef{ID: vRefV1.ID, Version: 99}),
+			data: orm.MarshalVersionedID(orm.VersionedIDRef{ID: vRefV1.ID, Version: 99}),
 		},
 		"By prefix query and ID": {
 			path:      "/electorates",
 			queryMode: weave.PrefixQueryMod,
-			data:      mustMarshal(t, &orm.VersionedIDRef{ID: vRefV1.ID}),
+			data:      vRefV1.ID,
 			exp:       []Electorate{electorateV1, electorateV2},
 		},
 		"By unknown ID": {
 			path: "/electorates",
-			data: mustMarshal(t, &orm.VersionedIDRef{ID: []byte{0x1}, Version: 1}),
+			data: orm.MarshalVersionedID(orm.VersionedIDRef{ID: []byte{0x1}, Version: 1}),
 		},
 		"By unknown key": {
 			path: "/electorates",